blacklist /opt and /srv
This is using a similar hack as d24b62d to blacklist /opt. By
referencing a non-existent file (/etc/firewarden/$now) we cause an empty
directory to be created. Firejail 0.9.46 adds --private-opt, so we might
as well use that instead of --whitelist. While we're at it, use the new
--private-srv too.

related to #1