drop --private-cache
This was useful for Firewarden pre-v1.1.0, where I didn't create a
private home unless the sandbox was being used to view a local file.
Since v1.1.0 Firewarden always creates a private home (to support the
configuration script option). So in effect we're just asking Firejail to
use a temporary filesystem for /home/user/.cache when /home/user is
already a temporary filesystem, which is redundant.

Closes #6