fix qutebrowser basedir arg for firejail 0.9.72
Something changed between Firejail 0.9.70 and 0.9.72 such that the old
method of providing the full `$dir/basedir` path caused Qutebrowser to
place its basedir into a private tmpfs, rather than placing its basedir
in the root of the tmpfs home directory (`$dir`) that was created for
the sandbox.

Since the application is being launched by Firejail using `$dir` as a
private home directory, simply passing the relative directory of
`basedir` (or an absolute path of `~/basedir`) has the same effect as
`$dir/basedir` did in Firejail versions up to 0.9.70.